Automatically backed up by Learn

+++ app.rb
@@ -14,7 +14,7 @@ class App < Sinatra::Base
   
   get '/say/:number/:phrase' do
     @phrase = params[:phrase]
-    params[:number].to_f.times do
+    params[:number].to_i.times do
       puts "#{@phrase}"
     end
   end
